+++ src/index.tsx
@@ -31,6 +31,7 @@ export { floatColumn } from './columns/floatColumn'
 export { intColumn } from './columns/intColumn'
 export { percentColumn } from './columns/percentColumn'
 export { dateColumn } from './columns/dateColumn'
+export { isoDateColumn } from './columns/isoDateColumn'
 export { keyColumn } from './columns/keyColumn'
 export { createAddRowsComponent } from './components/AddRows'
 export {